move check_unix_socket_path to ext/net lib.rs, ops_unix is unix-only but PipeWrap needs it on Windows too
