+++ ext/net/ops_unix.rs
@@ -29,7 +29,6 @@ use deno_core::RcRef;
 use deno_core::Resource;
 use deno_core::ResourceId;
 use deno_core::op2;
-use deno_permissions::CheckedPath;
 use deno_permissions::OpenAccessKind;
 use deno_permissions::PermissionsContainer;
 use serde::Deserialize;
@@ -38,7 +37,9 @@ use tokio::net::UnixDatagram;
 use tokio::net::UnixListener;
 pub use tokio::net::UnixStream;
 
+use crate::check_unix_socket_path;
 use crate::io::UnixStreamResource;
+use crate::is_unix_socket_abstract_path;
 use crate::ops::NetError;
 use crate::raw::NetworkListenerResource;
 
@@ -354,31 +355,6 @@ async fn send_to_unix_datagram(
   Ok(socket.send_to(buf, address_path).await?)
 }
 
-/// Permission check for opening a unix-domain socket path, shared with the
-/// node-compat `PipeWrap` path in `ext/node`.
-///
-/// Abstract socket paths (Linux, leading NUL) have no filesystem entry, so
-/// they skip the filesystem check and rely on the network check alone.
-pub fn check_unix_socket_path<'a>(
-  permissions: &mut PermissionsContainer,
-  path: Cow<'a, Path>,
-  access_kind: OpenAccessKind,
-  api_name: Option<&str>,
-) -> Result<CheckedPath<'a>, deno_permissions::PermissionCheckError> {
-  let checked = if is_unix_socket_abstract_path(path.as_ref()) {
-    CheckedPath::unsafe_new(path)
-  } else {
-    permissions.check_open(path, access_kind, api_name)?
-  };
-  // Unix sockets are an outbound network primitive, so require an
-  // `--allow-net=unix:<path>` rule in addition to filesystem access on the
-  // socket path. Without this, a script with only
-  // `--allow-read=/var/run/docker.sock` could connect to local IPC services
-  // (Docker, dbus, podman, etc.) with no `--allow-net` grant.
-  permissions.check_net_unix_socket(&checked, api_name)?;
-  Ok(checked)
-}
-
 #[cfg(any(target_os = "android", target_os = "linux"))]
 fn unix_sockaddr_from_abstract_path(
   path: &Path,
@@ -433,17 +409,4 @@ pub fn unix_socket_addr_path(
   }
 
   Ok(None)
-}
-
-fn is_unix_socket_abstract_path(path: &Path) -> bool {
-  #[cfg(any(target_os = "android", target_os = "linux"))]
-  {
-    path.as_os_str().as_bytes().first() == Some(&0)
-  }
-
-  #[cfg(not(any(target_os = "android", target_os = "linux")))]
-  {
-    let _ = path;
-    false
-  }
 }
